Version2

+++ src/main/java/com/foxpost/service/ClientService.java
@@ -7,6 +7,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -16,26 +18,47 @@ public class ClientService {
 
     public List<Clients> findByAllClients() {
         log.info(" <-- Find all Clients");
-//        log.error(" <-- Can't Find all Clients");
         return clientRepository.findAll();
     }
 
     public Clients getClientById(int id) {
         log.info(" <-- Find One Client by ID: {}", id);
-//        log.error(" <-- Cant Find One Client by ID: {}", id);
-        return clientRepository.findById(id).orElse(null);
+        Clients client = clientRepository.findById(id).orElse(null);
+        if (Objects.nonNull(client)) {
+            return client;
+        } else {
+            log.error(" <-- Cant Find One Client by ID: {}", id);
+            return null;
+        }
     }
+
     public void saveOrUpdate(Clients clients) {
-        log.info(" <-- in Save Client: {}", clients);
-//        log.error(" <-- Can't Save Client: {}", clients);
-        clientRepository.save(clients);
-        log.info(" <-- End Save Client: {}");
+        log.info(" <-- Save Client: {}", clients);
+        if (Objects.nonNull(clients)) {
+            clientRepository.save(clients);
+        } else {
+            log.error(" <-- Can't Save Client: {}", clients);
+        }
+        log.info(" <-- Save successfull: {} ", clients);
+    }
+
+    public Clients getParcelBySenderId(Clients senderId) {
+        Optional<Clients> parcel = clientRepository.findBySenderId(senderId);
+        if(parcel.isPresent()){
+            return parcel.get();
+        }else {
+            log.error("");
+            return null;
+        }
     }
     public void deleteClient(int id) {
         log.info(" <-- in Delete Client: {}", id);
-//        log.error(" <-- Can't Delete Client: {}", id);
-        clientRepository.deleteById(id);
-        log.info(" <-- End Delete Client: {}");
+        if (Objects.nonNull(id)) {
+            clientRepository.deleteById(id);
+        } else {
+            log.error(" <-- Can't Delete from Client: {}", id);
+        }
+        log.info("Delete successfull: {}", id);
     }
 
 }
