+++ src/main/java/com/foxpost/service/TrackingsService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @AllArgsConstructor
@@ -20,13 +21,27 @@ public class TrackingsService {
     private TrackingsRepository trackingsRepository;
 
     public List<Trackings> findAllTrackings() {
-        log.info("Find all Trackings");
-        return trackingsRepository.findAll();
+        log.info(" <-- Find all Trackings");
+      List  <Trackings> trackings = trackingsRepository.findAll();
+        if(Objects.nonNull(trackings)){
+            log.info(" <-- Find Successfull");
+            return trackingsRepository.findAll();
+        }else {
+            log.error(" <-- Not find ERROR");
+            return null;
+        }
+
     }
 
-    public List<Trackings> findAllTrackingsByParcelId(Parcels parcelId) {
-        log.info("Find all Trackings");
-        return trackingsRepository.findAllByParcelId(parcelsService.getParcelById(parcelId.getId()));
+    public List<Trackings> findAllTrackingsByParcelNumber(String parcelNumber) {
+        log.info(" <-- Find all Trackings");
+        Parcels parcel = parcelsService.getParcelByNumber(parcelNumber);
+        if (Objects.nonNull(parcel)) {
+            return trackingsRepository.findAllByParcelId(parcel);
+        } else {
+            log.error("Get in Tracks Parcel not Found");
+            return null;
+        }
     }
 
     public List<Parcels> getAllParcelsToOneTrack() {
@@ -34,27 +49,40 @@ public class TrackingsService {
     }
 
     public Trackings saveOrUpdate(TrackingDTO trackingDTO) {
-        log.info("in save Trackings --> {}", trackingDTO);
-        log.error("In save Trackings ERROR");
+        log.info(" <-- in save Trackings --> {}", trackingDTO);
         LocalDate eventDate = trackingDTO.getEventDate();
-        Trackings trackings = new Trackings(
+        Trackings tracking = new Trackings(
                 eventsService.getEventById(trackingDTO.getEventId()),
                 parcelsService.getParcelById(trackingDTO.getParcelId()),
                 eventDate
         );
-        log.info("save end Trackings --> {}", trackingDTO);
-        return trackingsRepository.save(trackings);
+        if (Objects.nonNull(tracking)) {
+            return trackingsRepository.save(tracking);
+        } else {
+            log.error("In save Trackings ERROR");
+            return null;
+        }
+
 
     }
 
     public void deleteTrackings(int id) {
-        log.info("delete from Trackings -> {}", id);
-        trackingsRepository.deleteById(id);
-        log.info("delete from Trackings END -> {}", id);
+        log.info(" <-- delete from Trackings -> {}", id);
+        if (Objects.nonNull(id)) {
+            trackingsRepository.deleteById(id);
+        }else {
+            log.error(" <-- Can't Delete from Trackings: {}", id);
+        }
     }
 
     public Trackings getTrackingById(int id) {
-        log.info("Tracking By ID -> {}", id);
-        return trackingsRepository.findById(id).orElse(null);
+        log.info(" <-- Find Tracking By ID -> {}", id);
+        Trackings track = trackingsRepository.findById(id).orElse(null);
+        if(Objects.nonNull(track)){
+            return track;
+        }else {
+            log.error(" <-- Can't Find track By Id: {}", id);
+            return null;
+        }
     }
 }
